# docs/openapi.yml
openapi: 3.0.3
info:
  title: Marketplace API
  version: 1.0.0
  description: Multi-vendor e-commerce platform API
paths:
...
  /api/catalog/products/:
    get:
      operationId: catalog_products_list
      parameters:
      - in: query
        name: brand
        schema:
          type: string
      - in: query
        name: category
        schema:
          type: string
      - in: query
        name: in_stock
        schema:
          type: boolean
      - in: query
        name: is_featured
        schema:
          type: boolean
      - in: query
        name: max_price
        schema:
          type: number
      - in: query
        name: min_price
        schema:
          type: number
      - in: query
        name: min_rating
        schema:
          type: number
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      - in: query
        name: seller
        schema:
          type: number
      - in: query
        name: status
        schema:
          type: string
          enum:
          - approved
          - draft
          - pending
          - rejected
        description: |-
          * `draft` - Draft
          * `pending` - Pending approval
          * `approved` - Approved
          * `rejected` - Rejected
      tags:
      - catalog
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedProductListList'
          description: ''
    post:
      operationId: catalog_products_create
      tags:
      - catalog
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductWriteRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ProductWriteRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ProductWriteRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductWrite'
          description: ''
...
components:
  schemas:
...
    PaginatedProductListList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/ProductList'
...
    ProductList:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 200
        slug:
          type: string
          maxLength: 220
          pattern: ^[-a-zA-Z0-9_]+$
        sku:
          type: string
          maxLength: 60
        price:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        compare_at_price:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          nullable: true
        discount_percent:
          type: integer
          readOnly: true
        thumbnail:
          type: string
          format: uri
          maxLength: 200
        rating:
          type: string
          format: decimal
          pattern: ^-?\d{0,1}(?:\.\d{0,2})?$
        review_count:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        stock:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        in_stock:
          type: boolean
          readOnly: true
        is_featured:
          type: boolean
        category_name:
          type: string
          readOnly: true
        brand_name:
          type: string
          readOnly: true
          default: ''
        store_name:
          type: string
          readOnly: true
          default: ''
        sold_count:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        status:
          $ref: '#/components/schemas/Status7f6Enum'
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - brand_name
      - category_name
      - created_at
      - discount_percent
      - id
      - in_stock
      - name
      - price
      - sku
      - store_name
...
    ProductWrite:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 200
        sku:
          type: string
          maxLength: 60
        category:
          type: integer
        brand:
          type: integer
          nullable: true
        short_description:
          type: string
          maxLength: 300
        description:
          type: string
        price:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        compare_at_price:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          nullable: true
        cost_price:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        tax_percent:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,2})?$
        stock:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        low_stock_threshold:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        weight_grams:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        thumbnail:
          type: string
          format: uri
          maxLength: 200
        is_active:
          type: boolean
        is_featured:
          type: boolean
        status:
          allOf:
          - $ref: '#/components/schemas/Status7f6Enum'
          readOnly: true
      required:
      - category
      - id
      - name
      - price
      - sku
      - status
    ProductWriteRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        sku:
          type: string
          minLength: 1
          maxLength: 60
        category:
          type: integer
        brand:
          type: integer
          nullable: true
        short_description:
          type: string
          maxLength: 300
        description:
          type: string
        price:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        compare_at_price:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          nullable: true
        cost_price:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        tax_percent:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,2})?$
        stock:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        low_stock_threshold:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        weight_grams:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        thumbnail:
          type: string
          format: uri
          maxLength: 200
        is_active:
          type: boolean
        is_featured:
          type: boolean
        image_urls:
          type: array
          items:
            type: string
            format: uri
            minLength: 1
          writeOnly: true
      required:
      - category
      - name
      - price
      - sku
...
    Status7f6Enum:
      enum:
      - draft
      - pending
      - approved
      - rejected
      type: string
      description: |-
        * `draft` - Draft
        * `pending` - Pending approval
        * `approved` - Approved
        * `rejected` - Rejected
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT